refactor(mountains): extract createTerrain into mountains/terrain-mesh.ts
Move createTerrain (the height-field -> PlaneGeometry rasterizer that pre-populates
the shared heightMap, applies the snow look, and scatters rocks + trees) into
src/mountains/terrain-mesh.ts. The mesh-vertex formula stays byte-identical to
getTerrainHeight in terrain.ts (the two-formula terrain contract); the mountains<->
trees import cycle still resolves at call time.

src/mountains/index.ts is now a slim assembly hub (~70 lines): it imports the five
submodules, builds the Mountains object, and re-exports the named/type surface. The
public ./mountains.js seam is unchanged.

Verified byte-identical: full Node suite green, physics-invariant harness OK, verify
49/49, tsc + lint clean, and `npm run build` succeeds.
